+++ app/controllers/agreements_controller.rb
@@ -144,7 +144,12 @@ class AgreementsController < ApplicationController
   end
 
   def print
-    render layout: false
+    file_pdf_location = Agreement.latex_file_location([@agreement], current_user)
+    if File.exists?(file_pdf_location)
+      send_file file_pdf_location, filename: "#{@agreement.user.last_name.gsub(/[^a-zA-Z\p{L}]/, '')}-#{@agreement.user.first_name.gsub(/[^a-zA-Z\p{L}]/, '')}-#{@agreement.agreement_number}-DAUA-#{(@agreement.submitted_at || @agreement.created_at).strftime("%Y-%m-%d")}.pdf", type: "application/pdf", disposition: "inline"
+    else
+      render layout: false
+    end
   end
 
   def download